refactor: extract trending podcast section to component
